pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "avcm"
version = "0.1.0"
description = "Autonomous Viral Content Machine — CrewAI multi-agent pipeline with ChromaDB memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.1.1",
    "crewai>=1.14.3",
    "crewai-tools>=1.14.3",
    "fastapi[standard]>=0.136.1",
    "openai>=2.33.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "typer>=0.23.1",
]

[project.optional-dependencies]
instagram = [
    "apify-client>=2.5.0",
]

[project.scripts]
avcm = "cli.main:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
    "respx>=0.23.1",
    "ruff>=0.15.12",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "tools", "memory", "api", "cli"]

[tool.pytest.ini_options]
testpaths = ["tests", "evals"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
packages = ["core", "agents", "tools", "memory", "api", "cli"]
